fix: resources crud implemented

+++ src/handlers/user.rs
@@ -1,12 +1,21 @@
 use std::sync::Arc;
 
 use crate::database::prelude::{Resource, ResourceGroup, User};
-use crate::database::{resource, resource::Model as ResourceModel, resource_group};
+use crate::database::{
+    resource,
+    resource::{ActiveModel as ResourceActiveModel, Model as ResourceModel},
+    resource_group, user,
+    user::{ActiveModel as UserActiveModel, Model as UserModel},
+};
+use crate::mappers::user::{AddResourceRequest, UpdateResourceRequest};
 use crate::mappers::DeleteResponse;
+use crate::utils::default_resource_checker::{is_default_resource, is_default_user};
 use axum::extract::Path;
 use axum::{Extension, Json};
+use chrono::Utc;
+use futures::future::try_join_all;
 use sea_orm::prelude::Uuid;
-use sea_orm::{ColumnTrait, EntityTrait, QueryFilter};
+use sea_orm::{ActiveModelTrait, ColumnTrait, EntityTrait, QueryFilter, Set};
 
 use crate::packages::db::AppState;
 use crate::{
@@ -17,13 +26,36 @@ use crate::{
     utils::role_checker::{is_current_realm_admin, is_master_realm_admin},
 };
 
-pub async fn get_users(Path(realm_id): Path<Uuid>) -> String {
-    format!("Hi from users of {realm_id}")
+pub async fn get_users(
+    user: TokenUser,
+    Extension(state): Extension<Arc<AppState>>,
+    Path(realm_id): Path<Uuid>,
+) -> Result<Json<Vec<UserModel>>, Error> {
+    if is_master_realm_admin(&user) || is_current_realm_admin(&user, &realm_id.to_string()) {
+        let users = User::find().filter(user::Column::RealmId.eq(realm_id)).all(&state.db).await?;
+        if users.is_empty() {
+            return Err(Error::not_found());
+        }
+        Ok(Json(users))
+    } else {
+        Err(Error::Authenticate(AuthenticateError::NoResource))
+    }
 }
 
-pub async fn get_user(Path((realm_id, user_id)): Path<(Uuid, Uuid)>) -> String {
-    println!("This is user Name: {} - {}", &realm_id, &user_id);
-    format!("user is - {} - {}", realm_id, user_id)
+pub async fn get_user(
+    user: TokenUser,
+    Extension(state): Extension<Arc<AppState>>,
+    Path((realm_id, user_id)): Path<(Uuid, Uuid)>,
+) -> Result<Json<UserModel>, Error> {
+    if is_master_realm_admin(&user) || is_current_realm_admin(&user, &realm_id.to_string()) {
+        let user = User::find_by_id(user_id).one(&state.db).await?;
+        match user {
+            Some(user) => Ok(Json(user)),
+            None => return Err(Error::Authenticate(AuthenticateError::NoResource)),
+        }
+    } else {
+        Err(Error::Authenticate(AuthenticateError::NoResource))
+    }
 }
 
 pub async fn delete_user(
@@ -32,6 +64,13 @@ pub async fn delete_user(
     Path((realm_id, user_id)): Path<(Uuid, Uuid)>,
 ) -> Result<Json<DeleteResponse>, Error> {
     if is_master_realm_admin(&user) || is_current_realm_admin(&user, &realm_id.to_string()) {
+        if is_default_user(user_id) {
+            return Err(Error::cannot_perform_operation("Cannot delete the default user"));
+        }
+        if user_id == user.sub {
+            return Err(Error::cannot_perform_operation("Cannot delete the current user"));
+        }
+
         let result = User::delete_by_id(user_id).exec(&state.db).await?;
         Ok(Json(DeleteResponse {
             ok: result.rows_affected == 1,
@@ -62,6 +101,125 @@ pub async fn get_resources(
             .all(&state.db)
             .await?;
         Ok(Json(resources))
+    } else {
+        Err(Error::Authenticate(AuthenticateError::ActionForbidden))
+    }
+}
+
+pub async fn add_resources(
+    user: TokenUser,
+    Extension(state): Extension<Arc<AppState>>,
+    Path((realm_id, user_id)): Path<(Uuid, Uuid)>,
+    Json(payload): Json<AddResourceRequest>,
+) -> Result<Json<Vec<ResourceModel>>, Error> {
+    if is_master_realm_admin(&user) || is_current_realm_admin(&user, &realm_id.to_string()) {
+        if payload.group_id.is_some() {
+            let futures: Vec<_> = payload
+                .identifiers
+                .iter()
+                .map(|(name, value)| {
+                    let resource = resource::ActiveModel {
+                        group_id: Set(payload.group_id.unwrap()),
+                        name: Set(name.to_string()),
+                        value: Set(value.to_string()),
+                        ..Default::default()
+                    };
+                    resource.insert(&state.db)
+                })
+                .collect();
+            let resources = try_join_all(futures).await?;
+            Ok(Json(resources))
+        } else if payload.group_name.is_some() {
+            let resource_groups = ResourceGroup::find()
+                .filter(resource_group::Column::RealmId.eq(realm_id))
+                .filter(resource_group::Column::UserId.eq(user_id))
+                .filter(resource_group::Column::Name.eq(payload.group_name))
+                .one(&state.db)
+                .await?;
+            if resource_groups.is_none() {
+                return Err(Error::not_found());
+            }
+            let resource_group = resource_groups.unwrap();
+
+            let futures: Vec<_> = payload
+                .identifiers
+                .iter()
+                .map(|(name, value)| {
+                    let resource = resource::ActiveModel {
+                        group_id: Set(resource_group.id),
+                        name: Set(name.to_string()),
+                        value: Set(value.to_string()),
+                        ..Default::default()
+                    };
+                    resource.insert(&state.db)
+                })
+                .collect();
+            let resources = try_join_all(futures).await?;
+            Ok(Json(resources))
+        } else {
+            Err(Error::cannot_perform_operation("Either group_name or group_id must be provided"))
+        }
+    } else {
+        Err(Error::Authenticate(AuthenticateError::ActionForbidden))
+    }
+}
+
+pub async fn update_resource(
+    user: TokenUser,
+    Extension(state): Extension<Arc<AppState>>,
+    Path((realm_id, _, resource_id)): Path<(Uuid, Uuid, Uuid)>,
+    Json(payload): Json<UpdateResourceRequest>,
+) -> Result<Json<ResourceModel>, Error> {
+    if is_master_realm_admin(&user) || is_current_realm_admin(&user, &realm_id.to_string()) {
+        if is_default_resource(resource_id) {
+            return Err(Error::cannot_perform_operation("Cannot update the default resource"));
+        }
+
+        let resource = Resource::find_by_id(resource_id).one(&state.db).await?;
+        if resource.is_none() {
+            return Err(Error::not_found());
+        }
+
+        let locked_at = match payload.lock {
+            Some(true) => Some(resource.as_ref().unwrap().locked_at.unwrap_or_else(|| Utc::now().naive_utc())),
+            Some(false) => None,
+            None => None,
+        };
+        let resource = ResourceActiveModel {
+            id: Set(resource_id),
+            group_id: Set(resource.unwrap().group_id),
+            name: Set(payload.name),
+            value: Set(payload.value),
+            description: Set(payload.description),
+            locked_at: Set(locked_at),
+            ..Default::default()
+        };
+        let resource = resource.update(&state.db).await?;
+        Ok(Json(resource))
+    } else {
+        Err(Error::Authenticate(AuthenticateError::ActionForbidden))
+    }
+}
+
+pub async fn delete_resource(
+    user: TokenUser,
+    Extension(state): Extension<Arc<AppState>>,
+    Path((realm_id, _, resource_id)): Path<(Uuid, Uuid, Uuid)>,
+) -> Result<Json<DeleteResponse>, Error> {
+    if is_master_realm_admin(&user) || is_current_realm_admin(&user, &realm_id.to_string()) {
+        if is_default_resource(resource_id) {
+            return Err(Error::cannot_perform_operation("Cannot delete the default resource"));
+        }
+
+        let resource = Resource::find_by_id(resource_id).one(&state.db).await?;
+        if resource.is_none() {
+            return Err(Error::not_found());
+        }
+
+        let result = Resource::delete_by_id(resource_id).exec(&state.db).await?;
+        Ok(Json(DeleteResponse {
+            ok: result.rows_affected == 1,
+        }))
     } else {
         Err(Error::Authenticate(AuthenticateError::ActionForbidden))
     }
